+++ frontend/src/components/Layout.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import { useState } from 'react';
 import SideNavbar from './navbar/SideNavbar';
 import { Outlet } from 'react-router-dom';
 import { Menu } from 'lucide-react';
@@ -23,16 +23,16 @@ const MainLayout = () => {
       {/* Menu Button */}
       <button
         onClick={toggleSidebar}
-        className={`fixed p-2 rounded-lg z-50 md:hidden ${
-          isSidebarVisible ? "text-white ml-64" : "text-primary ml-4"
-        } transition-all duration-300 ease-in-out`}
+        className={`fixed pt-8 rounded-lg z-50 md:hidden ${
+          isSidebarVisible ? "text-white" : "text-primary"
+        } ml-4`} 
       >
-        <Menu size={24} />
+        <Menu size={32} />
       </button>
 
       {/* Sidebar */}
       <aside
-        className={`fixed inset-y-0 left-0 w-64 bg-gray-800 text-white transform transition-transform duration-300 ease-in-out z-40 ${
+        className={`fixed inset-y-0 left-0 w-64 transform transition-transform duration-300 ease-in-out z-40 ${
           isSidebarVisible ? 'translate-x-0' : '-translate-x-full'
         } md:translate-x-0 md:relative`}
       >
